service_a/requirements.txt
fastapi
httpx
uvicorn[standard]
pyjwt